feat: implementação do datasource para salvar e obter usuarios
Este commit introduz um `UserRemoteDataSource` para interagir com o Firebase Firestore.

Principais alterações:
- Criado `UserRemoteDataSource` com métodos `saveUser` e `getUser` para salvar e recuperar dados do usuário do Firestore.
- `UserRepositoryImpl` agora delega as operações de usuário para o `UserRemoteDataSource`.
- A interface `UserRepository` foi atualizada para incluir o método `getUser`.
- Adicionado o campo `displayName` ao modelo `User`.
