+++ app/src/main/java/com/example/capdex/data/repository/UserRepository.kt
@@ -4,4 +4,5 @@ import com.example.capdex.domain.model.User
 
 interface UserRepository {
     suspend fun saveUser(user: User): Result<Unit>
+    suspend fun getUser(uid: String): Result<User?>
 }
