+++ app/src/main/java/com/example/capdex/data/repository/UserRepositoryImpl.kt
@@ -1,23 +1,17 @@
 package com.example.capdex.data.repository
 
+import com.example.capdex.data.remote.UserRemoteDataSource
 import com.example.capdex.domain.model.User
-import com.google.firebase.firestore.FirebaseFirestore
-import kotlinx.coroutines.tasks.await
 import javax.inject.Inject
 
 class UserRepositoryImpl @Inject constructor(
-    private val firestore: FirebaseFirestore
+    private val remoteDataSource: UserRemoteDataSource
 ) : UserRepository {
-    private val usersCollection = firestore.collection("users")
-
     override suspend fun saveUser(user: User): Result<Unit> {
-        return try {
-            usersCollection.document(user.uid)
-                .set(user)
-                .await()
-            Result.success(Unit)
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
+        return remoteDataSource.saveUser(user)
+    }
+
+    override suspend fun getUser(uid: String): Result<User?> {
+        return remoteDataSource.getUser(uid) // Chamando a função correspondente no DataSource
     }
 }
