+++ app/src/main/java/com/example/capdex/domain/model/User.kt
@@ -4,4 +4,5 @@ data class User(
     val uid: String = "",
     val email: String? = null,
     val userType: String = "comum",
+    val displayName: String? = null,
 )
